+++ examples/cat_test.ts
@@ -6,7 +6,7 @@ const moduleDir = dirname(fromFileUrl(import.meta.url));
 
 Deno.test("[examples/cat] print multiple files", async () => {
   const decoder = new TextDecoder();
-  const { stdout } = await Deno.spawn(Deno.execPath(), {
+  const command = new Deno.Command(Deno.execPath(), {
     args: [
       "run",
       "--quiet",
@@ -18,6 +18,7 @@ Deno.test("[examples/cat] print multiple files", async () => {
     cwd: moduleDir,
     stdout: "piped",
   });
+  const { stdout } = await command.output();
 
   try {
     const actual = decoder.decode(stdout).trim();