+++ fs/expand_glob.ts
@@ -9,7 +9,7 @@ import {
   SEP_PATTERN,
 } from "../path/mod.ts";
 import { walk, walkSync } from "./walk.ts";
-import { assert } from "../_util/assert.ts";
+import { assert } from "../_util/asserts.ts";
 import { isWindows } from "../_util/os.ts";
 import {
   createWalkEntry,
@@ -91,11 +91,15 @@ export async function* expandGlob(
     .map((s: string): RegExp => globToRegExp(s, globOptions));
   const shouldInclude = (path: string): boolean =>
     !excludePatterns.some((p: RegExp): boolean => !!path.match(p));
-  const { segments, isAbsolute: isGlobAbsolute, hasTrailingSep, winRoot } =
-    split(toPathString(glob));
+  const {
+    segments,
+    isAbsolute: isGlobAbsolute,
+    hasTrailingSep,
+    winRoot,
+  } = split(toPathString(glob));
 
   let fixedRoot = isGlobAbsolute
-    ? (winRoot != undefined ? winRoot : "/")
+    ? winRoot != undefined ? winRoot : "/"
     : absRoot;
   while (segments.length > 0 && !isGlob(segments[0])) {
     const seg = segments.shift();
@@ -127,7 +131,10 @@ export async function* expandGlob(
       }
       return;
     } else if (globSegment == "**") {
-      return yield* walk(walkInfo.path, { skip: excludePatterns });
+      return yield* walk(walkInfo.path, {
+        skip: excludePatterns,
+        maxDepth: globstar ? Infinity : 1,
+      });
     }
     const globPattern = globToRegExp(globSegment, globOptions);
     for await (
@@ -137,7 +144,8 @@ export async function* expandGlob(
       })
     ) {
       if (
-        walkEntry.path != walkInfo.path && walkEntry.name.match(globPattern)
+        walkEntry.path != walkInfo.path &&
+        walkEntry.name.match(globPattern)
       ) {
         yield walkEntry;
       }
@@ -149,13 +157,16 @@ export async function* expandGlob(
     // Advancing the list of current matches may introduce duplicates, so we
     // pass everything through this Map.
     const nextMatchMap: Map<string, WalkEntry> = new Map();
-    await Promise.all(currentMatches.map(async (currentMatch) => {
-      for await (const nextMatch of advanceMatch(currentMatch, segment)) {
-        nextMatchMap.set(nextMatch.path, nextMatch);
-      }
-    }));
+    await Promise.all(
+      currentMatches.map(async (currentMatch) => {
+        for await (const nextMatch of advanceMatch(currentMatch, segment)) {
+          nextMatchMap.set(nextMatch.path, nextMatch);
+        }
+      }),
+    );
     currentMatches = [...nextMatchMap.values()].sort(comparePath);
   }
+
   if (hasTrailingSep) {
     currentMatches = currentMatches.filter(
       (entry: WalkEntry): boolean => entry.isDirectory,
@@ -199,11 +210,15 @@ export function* expandGlobSync(
     .map((s: string): RegExp => globToRegExp(s, globOptions));
   const shouldInclude = (path: string): boolean =>
     !excludePatterns.some((p: RegExp): boolean => !!path.match(p));
-  const { segments, isAbsolute: isGlobAbsolute, hasTrailingSep, winRoot } =
-    split(toPathString(glob));
+  const {
+    segments,
+    isAbsolute: isGlobAbsolute,
+    hasTrailingSep,
+    winRoot,
+  } = split(toPathString(glob));
 
   let fixedRoot = isGlobAbsolute
-    ? (winRoot != undefined ? winRoot : "/")
+    ? winRoot != undefined ? winRoot : "/"
     : absRoot;
   while (segments.length > 0 && !isGlob(segments[0])) {
     const seg = segments.shift();
@@ -235,7 +250,10 @@ export function* expandGlobSync(
       }
       return;
     } else if (globSegment == "**") {
-      return yield* walkSync(walkInfo.path, { skip: excludePatterns });
+      return yield* walkSync(walkInfo.path, {
+        skip: excludePatterns,
+        maxDepth: globstar ? Infinity : 1,
+      });
     }
     const globPattern = globToRegExp(globSegment, globOptions);
     for (
@@ -245,7 +263,8 @@ export function* expandGlobSync(
       })
     ) {
       if (
-        walkEntry.path != walkInfo.path && walkEntry.name.match(globPattern)
+        walkEntry.path != walkInfo.path &&
+        walkEntry.name.match(globPattern)
       ) {
         yield walkEntry;
       }
@@ -264,6 +283,7 @@ export function* expandGlobSync(
     }
     currentMatches = [...nextMatchMap.values()].sort(comparePath);
   }
+
   if (hasTrailingSep) {
     currentMatches = currentMatches.filter(
       (entry: WalkEntry): boolean => entry.isDirectory,