+++ node/_core.ts
@@ -22,8 +22,18 @@ if (Deno?.core) {
     encode(chunk: string): Uint8Array {
       return new TextEncoder().encode(chunk);
     },
+    decode(chunk: Uint8Array): string {
+      return new TextDecoder().decode(chunk);
+    },
     eventLoopHasMoreWork(): boolean {
       return false;
+    },
+    ops: {
+      op_napi_open(_filename: string) {
+        throw new Error(
+          "Node API is not supported in this environment",
+        );
+      },
     },
   };
 }