+++ node/_process/streams.mjs
@@ -41,17 +41,13 @@ function createWritableStdioStream(writer, name) {
       enumerable: true,
       configurable: true,
       get: () =>
-        Deno.isatty?.(writer?.rid)
-          ? Deno.consoleSize?.(writer?.rid).columns
-          : undefined,
+        Deno.isatty?.(writer?.rid) ? Deno.consoleSize?.().columns : undefined,
     },
     rows: {
       enumerable: true,
       configurable: true,
       get: () =>
-        Deno.isatty?.(writer?.rid)
-          ? Deno.consoleSize?.(writer?.rid).rows
-          : undefined,
+        Deno.isatty?.(writer?.rid) ? Deno.consoleSize?.().rows : undefined,
     },
     isTTY: {
       enumerable: true,
@@ -63,7 +59,7 @@ function createWritableStdioStream(writer, name) {
       configurable: true,
       value: () =>
         Deno.isatty?.(writer?.rid)
-          ? Object.values(Deno.consoleSize?.(writer?.rid))
+          ? Object.values(Deno.consoleSize?.())
           : undefined,
     },
   });