+++ node/http.ts
@@ -57,6 +57,12 @@ const METHODS = [
 
 type Chunk = string | Buffer | Uint8Array;
 
+// @ts-ignore Deno[Deno.internal] is used on purpose here
+const DenoServe = Deno[Deno.internal]?.nodeUnstable?.serve || Deno.serve;
+// @ts-ignore Deno[Deno.internal] is used on purpose here
+const DenoUpgradeHttpRaw = Deno[Deno.internal]?.nodeUnstable?.upgradeHttpRaw ||
+  Deno.upgradeHttpRaw;
+
 function chunkToU8(chunk: Chunk): Uint8Array {
   if (typeof chunk === "string") {
     return core.encode(chunk);
@@ -180,6 +186,10 @@ class ClientRequest extends NodeWritable {
       port === 80 ? "" : `:${port}`
     }${path}`;
   }
+
+  setTimeout() {
+    console.log("not implemented: ClientRequest.setTimeout");
+  }
 }
 
 /** IncomingMessage for http(s) client */
@@ -233,6 +243,9 @@ export class ServerResponse extends NodeWritable {
   statusMessage?: string = undefined;
   #headers = new Headers({});
   #readable: ReadableStream;
+  override writable = true;
+  // used by `npm:on-finished`
+  finished = false;
   headersSent = false;
   #firstChunk: Chunk | null = null;
   // Used if --unstable flag IS NOT present
@@ -337,7 +350,7 @@ export class ServerResponse extends NodeWritable {
     const body = singleChunk ?? (final ? null : this.#readable);
     if (this.#isFlashRequest) {
       this.#resolve!(
-        new Response(Buffer.isBuffer(body) ? body.toString() : body, {
+        new Response(body, {
           headers: this.#headers,
           status: this.statusCode,
           statusText: this.statusMessage,
@@ -358,6 +371,7 @@ export class ServerResponse extends NodeWritable {
 
   // deno-lint-ignore no-explicit-any
   override end(chunk?: any, encoding?: any, cb?: any): this {
+    this.finished = true;
     if (this.#isFlashRequest) {
       // Flash sets both of these headers.
       this.#headers.delete("transfer-encoding");
@@ -454,7 +468,7 @@ class ServerImpl extends EventEmitter {
   constructor(handler?: ServerHandler) {
     super();
     // @ts-ignore Might be undefined without `--unstable` flag
-    this.#isFlashServer = typeof Deno.serve == "function";
+    this.#isFlashServer = typeof DenoServe == "function";
     if (this.#isFlashServer) {
       this.#servePromise = deferred();
       this.#servePromise.then(() => this.emit("close"));
@@ -552,7 +566,7 @@ class ServerImpl extends EventEmitter {
     const handler = (request: Request) => {
       const req = new IncomingMessageForServer(request);
       if (req.upgrade && this.listenerCount("upgrade") > 0) {
-        const [conn, head] = Deno.upgradeHttpRaw(request) as [
+        const [conn, head] = DenoUpgradeHttpRaw(request) as [
           Deno.Conn,
           Uint8Array,
         ];
@@ -572,7 +586,7 @@ class ServerImpl extends EventEmitter {
       return;
     }
     this.#ac = ac;
-    Deno.serve(
+    DenoServe(
       {
         handler: handler as Deno.ServeHandler,
         ...this.#addr,