+++ node/module.ts
@@ -27,7 +27,7 @@ import nodeMods from "./module_all.ts";
 import upstreamMods from "./upstream_modules.ts";
 
 import * as path from "../path/mod.ts";
-import { assert } from "../_util/assert.ts";
+import { assert } from "../_util/asserts.ts";
 import { fileURLToPath, pathToFileURL } from "./url.ts";
 import { isWindows } from "../_util/os.ts";
 import {
@@ -1350,7 +1350,9 @@ Module._extensions[".json"] = (module: Module, filename: string) => {
   }
 };
 
-// .node extension is not supported
+Module._extensions[".node"] = (module: Module, filename: string) => {
+  module.exports = core.ops.op_napi_open(filename);
+};
 
 function createRequireFromPath(filename: string): RequireFunction {
   // Allow a directory to be passed as the filename